- switch to prefers-color-scheme to support all major platforms - remove typescript - rename package
