--- lib/tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "compilerOptions": {
-        "forceConsistentCasingInFileNames": true,
-        "module": "commonjs",
-        "noUnusedLocals": true,
-        "noUnusedParameters": true,
-        "outDir": "../dist",
-        "preserveConstEnums": true,
-        "removeComments": true,
-        "target": "es2017",
-        "typeRoots": [
-            "../node_modules/@types"
-        ]
-    },
-    "compileOnSave": true
-}
